move collection serializers into collection pkg
