+++ java/fury-core/src/main/java/io/fury/serializer/CompatibleSerializer.java
@@ -23,6 +23,7 @@ import io.fury.resolver.ClassInfo;
 import io.fury.resolver.ClassResolver;
 import io.fury.resolver.FieldResolver;
 import io.fury.resolver.RefResolver;
+import io.fury.serializer.collection.CollectionSerializer;
 import io.fury.util.FieldAccessor;
 import io.fury.util.Platform;
 import io.fury.util.ReflectionUtils;
@@ -237,8 +238,8 @@ public final class CompatibleSerializer<T> extends CompatibleSerializerBase<T> {
     // following write is consistent with `BaseSeqCodecBuilder.serializeForCollection`
     ClassInfo classInfo = fieldInfo.getClassInfo(fieldValue.getClass());
     classResolver.writeClass(buffer, classInfo);
-    CollectionSerializers.CollectionSerializer collectionSerializer =
-        (CollectionSerializers.CollectionSerializer) classInfo.getSerializer();
+    CollectionSerializer collectionSerializer =
+        (CollectionSerializer) classInfo.getSerializer();
     collectionSerializer.setElementSerializer(elementClassInfo.getSerializer());
     collectionSerializer.write(buffer, fieldValue);
   }
