+++ java/fury-core/src/main/java/io/fury/serializer/SynchronizedSerializers.java
@@ -19,7 +19,7 @@ package io.fury.serializer;
 import io.fury.Fury;
 import io.fury.collection.Tuple2;
 import io.fury.memory.MemoryBuffer;
-import io.fury.serializer.CollectionSerializers.CollectionSerializer;
+import io.fury.serializer.collection.CollectionSerializer;
 import io.fury.serializer.MapSerializers.MapSerializer;
 import io.fury.util.LoggerFactory;
 import io.fury.util.Platform;
